refactor POE eligibility validation | organize directory
