+++ src/applications/edu-benefits/10275/hooks/useValidateFacilityCode.js
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { apiRequest } from 'platform/utilities/api';
 import { setData } from 'platform/forms-system/src/js/actions';
+import { isPOEEligible } from '../helpers';
 
 export const useValidateFacilityCode = formData => {
   const [loader, setLoader] = useState(false);
@@ -24,12 +25,6 @@ export const useValidateFacilityCode = formData => {
           );
           const attrs = response.data.attributes;
 
-          const programTypes = Array.isArray(attrs.programTypes)
-            ? attrs.programTypes
-            : [];
-          const eligible =
-            programTypes.includes('IHL') || programTypes.includes('NCD');
-
           const institutionAddress = {
             street: attrs.address1 || '',
             street2: attrs.address2 || '',
@@ -47,7 +42,7 @@ export const useValidateFacilityCode = formData => {
                 ...formData.institutionDetails,
                 institutionName: response?.data?.attributes?.name,
                 institutionAddress,
-                poeEligible: eligible,
+                poeEligible: isPOEEligible(facilityCode),
               },
             }),
           );
