+++ src/applications/edu-benefits/10275/tests/containers/InstitutionAddress.unit.spec.jsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import { expect } from 'chai';
-import InstitutionAddress from '../../components/InstitutionAddress';
+import InstitutionAddress from '../../containers/InstitutionAddress';
 
 const mockStore = configureStore([]);
